Prompt: try prompting even without extracted window ID
For better compatiblity with other Secret Service providers than Gnome
Keyring, attempt to call prompt even if window ID cannot be extracted
from the prompt ID.

Storing window ID inside prompt ID doesn't seem to be a part of
secret-service specification, perhaps it is Gnome-specific and other
services does not guarnatee returning it in the same format.

I have tried this patch with KeePassXC and calling a prompt with empty
window ID seem to work just fine, as opposed to not calling it at all,
which creates uncalled prompts in DBus and makes code hangs until
timeout is hit.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>